+++ react/components/Newsletter/index.js
@@ -5,6 +5,7 @@ import { injectIntl, intlShape } from 'react-intl'
 import { Input, Button } from 'vtex.styleguide'
 import SUBSCRIBE_NEWSLETTER from './mutations/subscribeNewsletter.graphql'
 import style from './style.css'
+import { formatIOMessage } from 'vtex.native-types'
 
 const EMAIL_REGEX = /^[A-z0-9+_-]+(?:\.[A-z0-9+_-]+)*@(?:[A-z0-9](?:[A-z0-9-]*[A-z0-9])?\.)+[A-z0-9](?:[A-z0-9-]*[A-z0-9])?$/
 
@@ -69,14 +70,10 @@ class Newsletter extends Component {
   }
 
   render() {
-    const {
-      placeholder = this.props.intl.formatMessage({
-        id: 'store/newsletter.placeholder',
-      }),
-      submit = this.props.intl.formatMessage({ id: 'store/newsletter.submit' }),
-      label = this.props.intl.formatMessage({ id: 'store/newsletter.label' }),
-      hideLabel,
-    } = this.props
+    const { hideLabel, intl, submit, label, placeholder } = this.props
+    const submitText = formatIOMessage({ id: submit, intl })
+    const labelText = formatIOMessage({ id: label, intl })
+    const placeholderText = formatIOMessage({ id: placeholder, intl })
 
     return (
       <div
@@ -88,13 +85,15 @@ class Newsletter extends Component {
           {this.state.success ? (
             <Fragment>
               <div className={`${style.confirmationTitle} t-heading-3 pb4 tc`}>
-                {this.props.intl.formatMessage({
+                {formatIOMessage({
                   id: 'store/newsletter.confirmationTitle',
+                  intl,
                 })}
               </div>
               <div className={`${style.confirmationText} t-body tc`}>
-                {this.props.intl.formatMessage({
+                {formatIOMessage({
                   id: 'store/newsletter.confirmationText',
+                  intl,
                 })}
               </div>
             </Fragment>
@@ -106,7 +105,7 @@ class Newsletter extends Component {
                 }`}
                 htmlFor="newsletter-input"
               >
-                {label}
+                {labelText}
               </label>
               <div className={`${style.inputGroup} flex-ns pt5`}>
                 <Input
@@ -114,12 +113,13 @@ class Newsletter extends Component {
                   id="newsletter-input"
                   errorMessage={
                     this.state.invalidEmail
-                      ? this.props.intl.formatMessage({
+                      ? formatIOMessage({
                           id: 'store/newsletter.invalidEmail',
+                          intl,
                         })
                       : null
                   }
-                  placeholder={placeholder}
+                  placeholder={placeholderText}
                   name="newsletter"
                   value={this.state.email}
                   onChange={this.handleChangeEmail}
@@ -135,13 +135,16 @@ class Newsletter extends Component {
                     onClick={this.handleSubmit}
                     isLoading={this.state.loading}
                   >
-                    {submit}
+                    {submitText}
                   </Button>
                 </div>
               </div>
               {this.state.error && (
                 <div className={`${style.error} c-danger t-body pt5`}>
-                  {this.props.intl.formatMessage({ id: 'store/newsletter.error' })}
+                  {formatIOMessage({
+                    id: 'store/newsletter.error',
+                    intl,
+                  })}
                 </div>
               )}
             </form>
@@ -178,21 +181,6 @@ Newsletter.getSchema = () => {
         title: 'admin/editor.newsletter.hideLabel',
         default: false,
         isLayout: true,
-      },
-      label: {
-        type: 'string',
-        title: 'admin/editor.newsletter.label',
-        isLayout: false,
-      },
-      placeholder: {
-        type: 'string',
-        title: 'admin/editor.newsletter.placeholder',
-        isLayout: false,
-      },
-      submit: {
-        type: 'string',
-        title: 'admin/editor.newsletter.submit',
-        isLayout: false,
       },
     },
   }